Add back ERL_LDFLAGS
In release 0.1.9, I had to pin the rebar3 plugin port_compiler
at version v1.10.2 in order to make it compile in windows.
ERL_LDFLAGS was removed for no good reason.
In order to make it work with OTP-23, we need to add it
back because port_compiler v1.10.2 does not have the fix for OTP-23